[test.products_boundaryl2/elliptic] fix unsafe conversion warning

+++ test/products_weightedl2.hh
@@ -6,6 +6,8 @@
 #ifndef DUNE_GDT_TEST_PRODUCTS_WEIGHTEDL2_HH
 #define DUNE_GDT_TEST_PRODUCTS_WEIGHTEDL2_HH
 
+#include <boost/numeric/conversion/cast.hpp>
+
 #include <dune/stuff/la/container/common.hh>
 
 #include <dune/gdt/discretefunction/default.hh>
@@ -33,7 +35,7 @@ struct WeightedL2ProductBase
   typedef Stuff::Functions::Expression< EntityType, DomainFieldType, dimDomain, RangeFieldType, dimRange > FunctionType;
 
   WeightedL2ProductBase()
-   : grid_(GridProviderType(0.0, 1.0, dsc_grid_elements()).grid_ptr())
+   : grid_(GridProviderType(0.0, 1.0, boost::numeric_cast< unsigned int >(dsc_grid_elements())).grid_ptr())
    , space_(Dune::GDT::SpaceTools::GridPartView< SpaceType >::create_leaf(*grid_))
    , one_("x", "1.0", 0)
   {}
